refactor: establish code quality foundations
